% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration_fit.R, R/configuration_fit_set.R
\name{configuration_fit}
\alias{configuration_fit}
\alias{is.configuration_fit}
\alias{configuration_fit_set}
\alias{is.configuration_fit_set}
\title{Configuration fit and fit set objects}
\usage{
is.configuration_fit(x)

is.configuration_fit_set(x)
}
\value{
A \code{configuration_fit} object has the following elements:
\itemize{
\item{\code{x}: The group network used to fit a configuration.}
\item{\code{fit}: The best-fitting configuration with its rows and columns
reordered to correspond with its best fit to the group network.}
\item{\code{configuration_id}: The \code{id} associated with the
configuration.}
\item{\code{score}: The fit score for the configuration.}
\item{\code{potential}: The highest (or lowest when \code{maximize = FALSE})
potential score given \code{x}.}
}
A \code{configuration_fit} object also has the following attributes:
\itemize{
\item{\code{duration}: Duration of solver function.}
\item{\code{solver}: Name of solver used.}
\item{\code{ROI_obj}: Returned object from the \code{\link{ROI_solve}}
function, if used.}
\item{\code{lp_structure}: Structure of the linear program submitted to
\code{ROI_solve} function, if used.}
}
}
\description{
A \code{configuration_fit} object is returned by the \code{\link{
fit_configuration}} function for a given group network. A \code{
configuration_fit_set} object is returned by the \code{\link{
fit_configuration_set}} function when \code{ties.method = "all"}.
}
\seealso{
\code{\link{fit_configuration}}, \code{\link{fit_configuration_set}}
}
