% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration_set.R
\name{configuration_set}
\alias{configuration_set}
\alias{as.configuration_set}
\alias{is.configuration_set}
\alias{set_configuration_ids}
\title{Configuration set}
\usage{
configuration_set(...)

as.configuration_set(x)

is.configuration_set(x)

set_configuration_ids(x, ids = seq_along(x), ...)
}
\arguments{
\item{...}{\code{configuration} objects, \code{configuration_set} objects, matrices,
edge lists, or other objects that can be coerced to \code{configuration} objects.}

\item{x}{A list of \code{configuration} objects, matrices, edge lists, or other objects
that can be coerced to \code{configuration} objects.}

\item{ids}{A vector of IDs of the same length as \code{x}.}
}
\value{
A \code{configuration_set}.
}
\description{
\code{configuration_set} creates a \code{configuration_set} object, which is
a list of configurations.
}
\details{
A configuration set is fitted to one or more group networks. See
\code{\link{fit_group_network}} for details.

\code{set_configuration_ids} creates a \code{configuration_set} and sets the
configurations' IDs.
}
\examples{
f_set <- star(2:5)
get_attribute(f_set, "id")
f_set <- set_configuration_ids(f_set, LETTERS[1:4])
get_attribute(f_set, "id")
}
\seealso{
\code{\link{fit_group_network}}, \code{\link{configuration}}
}
