% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute_methods.R
\name{filter_by_attribute}
\alias{filter_by_attribute}
\title{Filter by attribute}
\usage{
filter_by_attribute(x, attribute, as.boolean = FALSE, ...)
}
\arguments{
\item{x}{A \code{configuration_set} or \code{configuration_fit_set}.}

\item{attribute}{A named list containing the filter criteria. Each element in the list represents
an attribute to filter on. A list element's name should be an attribute name.
The element value should include one or more attribute values to select
\code{configuration} or  \code{configuration_fit} objects. To be selected, a
\code{configuration} or \code{configuration_fit} object must match with each
attribute in the filter.}

\item{as.boolean}{When \code{TRUE}, the function returns a boolean vector for whether each \code{
configuration} or \code{configuration_fit} object matched the attribute filter.
When \code{FALSE}, the function returns the matched objects.}
}
\value{
A subset of objects that match the filter criteria or a boolean vector that
can be used to filter a \code{configuration_set} or \code{configuration_fit_set},
depending on the \code{as.boolean} parameter.
}
\description{
Filters a \code{configuration_set} or \code{configuration_fit_set} by one
or more attributes' values.
}
\examples{
f_set <- c(star(2), star(3), star(4))
filter_by_attribute(f_set, list(group_size = 2:3))
f_set[filter_by_attribute(f_set, list(group_size = 2:3), as.boolean = TRUE)]
}
\seealso{
\code{\link{get_attribute}}, \code{\link{configuration_set}},
\code{\link{configuration_fit_set}}
}
