% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_configuration_set.R
\name{fit_configuration_set}
\alias{fit_configuration_set}
\title{Fit configuration to group network}
\usage{
fit_configuration_set(x, configuration_set, ...)
}
\arguments{
\item{x}{A group network or list of group networks. A group network can be an adjacency
matrix, edge list, \code{igraph} object, or \code{network} object. \code{x} may
also be an edge list for multiple groups when \code{group_index} is provided.}

\item{configuration}{A \code{configuration} object.}
}
\value{
A \code{configuration_fit} object.
}
\description{
\code{fit_configuration} fits a configuration to group networks and returns
a \code{configuration_fit} object.
}
\details{
\code{fit_configuration} fits a single configuration. To fit multiple
configurations to group networks, see \code{\link{fit_configuration_set}}.
}
