% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration_fit.R
\name{is.configuration_fit}
\alias{is.configuration_fit}
\title{Configuration fit}
\usage{
is.configuration_fit(x)
}
\arguments{
\item{x}{An object.}
}
\value{
\code{TRUE} if a \code{configuration_fit} object; \code{FALSE} otherwise.
}
\description{
Returns whether an object is a \code{configuration_fit} object.
}
\seealso{
\code{\link{fit_group_network}}
}
