% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component.R
\name{ring}
\alias{ring}
\title{Create a ring configuration}
\usage{
ring(
  n,
  group_size = max(n),
  mode = c("mutual", "in", "out"),
  loops = FALSE,
  value = 1L
)
}
\arguments{
\item{n}{Vector for the number of connected vertices.}

\item{group_size}{Size of the group.}

\item{mode}{Can be \code{"mutual"}, \code{"in"}, or \code{"out"}: for edge \emph{(i, j)} and matrix \emph{M},
out creates edge \code{M[i, j]}, in creates edge \code{M[j, i]}, and mutual creates both.}

\item{value}{Edge value (or weight) to set.}
}
\value{
A list of group \code{configuration}s.
}
\description{
\code{ring} creates one or more group configurations in which vertices are serially connected.
}
\examples{
ring(4)
ring(2:4, 6)
}
