% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SmallGroupNetwork-package.R
\docType{package}
\name{SmallGroupNetwork}
\alias{SmallGroupNetwork}
\title{SmallGroupNetwork}
\description{
Fit group configurations to a group network to determine the best-fitting
configuration. See \code{\link{fit_group_network}} and \code{\link{configuration}}.
}
