% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component.R
\name{star}
\alias{star}
\title{Create a star configuration}
\usage{
star(
  n,
  group_size = max(n),
  mode = c("mutual", "in", "out"),
  loops = FALSE,
  value = 1L
)
}
\arguments{
\item{n}{Vector for the number of connected vertices.}

\item{group_size}{Size of the group.}

\item{mode}{Can be \code{"mutual"}, \code{"in"}, or \code{"out"}: for edge \emph{(i, j)} and matrix \emph{M},
\code{"out"} creates edge \code{M[i, j]}, \code{"in"} creates edge \code{M[j, i]}, and \code{"mutual"} creates both.}

\item{value}{Edge value (or weight) to set.}
}
\value{
A list of group \code{configuration}s.
}
\description{
\code{star} creates one or more group configurations in which \emph{n - 1} vertices are connected
to the first vertex.
}
\examples{
star(4)
star(2:4, 6)
}
