% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component.R
\name{subgroup}
\alias{subgroup}
\alias{subgroup_all}
\title{Create a subgroup configuration}
\usage{
subgroup(
  ...,
  group_size = NULL,
  mode = c("mutual", "in", "out"),
  relation = c("within", "between"),
  loops = FALSE,
  value = 1L
)

subgroup_all(
  n,
  group_size = max(n),
  min_size = 2,
  mode = c("mutual", "in", "out"),
  relation = c("within", "between"),
  loops = FALSE,
  value = 1L
)
}
\arguments{
\item{...}{One or more vectors of subgroup sizes. If \code{relation = "between"}, at least two numbers
or vectors are required.}

\item{group_size}{Size of the group.}

\item{mode}{Can be \code{"mutual"}, \code{"in"}, or \code{"out"}: for edge \emph{(i, j)} and matrix \emph{M},
out creates edge \code{M[i, j]}, in creates edge \code{M[j, i]}, and mutual creates both.}

\item{relation}{Specifies whether edges are \code{"within"} or \code{"between"} subgroups. Default is
\code{"within"}.}

\item{value}{Edge value (or weight) to set.}

\item{n}{Vector for the number of vertices in subgroups.}

\item{min_size}{Smallest subgroup size allowed.}
}
\value{
A list of group \code{configuration}s.
}
\description{
\code{ring} creates one or more group configurations in which vertices are connected within or
between subgroups.
}
\details{
If the vector lengths in \code{...} are greater than one, then multiple configurations will be
created. Within a group, subgroups are sorted in decreasing size order.
}
\section{Functions}{
\itemize{
\item \code{subgroup}: Create subgroup combinations defined in \code{...}.

\item \code{subgroup_all}: Create all possible subgroup combinations.
}}

\examples{
subgroup(4)
subgroup(2, 4)
subgroup(2:4)
subgroup(2, 2:4, 3)
subgroup(2, 4, group_size = 8, relation = "between")

subgroup_all(6)
subgroup_all(6, relation = "between")
subgroup_all(6, group_size = 8)
}
