% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration_fit_set.R
\name{configuration_fit_set}
\alias{configuration_fit_set}
\alias{is.configuration_fit_set}
\title{Configuration fit set}
\usage{
configuration_fit_set(...)

is.configuration_fit_set(x)
}
\value{
A \code{configuration_fit_set}.
}
\description{
\code{configuration_fit_set} is a list of \code{configuration_fit} objects
returned by the \code{fit_configuration_set} function when \code{
ties.method = "all"}.
}
\seealso{
\code{\link{fit_configuration_set}}, \code{\link{configuration_fit}}
}
