% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration_set.R
\name{configuration_set}
\alias{configuration_set}
\alias{set_configuration_ids}
\alias{as.configuration_set}
\title{Configuration set}
\usage{
configuration_set(...)

set_configuration_ids(x, ids = seq_along(x), ...)

as.configuration_set(x, ...)
}
\arguments{
\item{...}{\code{configuration} objects, \code{configuration_set} objects, matrices, edge lists, or
other objects that can be coerced to \code{configuration} objects
(see \code{\link{configuration}}).}

\item{x}{A list of \code{configuration} objects, matrices, edge lists, or other objects that can be
coerced to \code{configuration} objects.}

\item{ids}{A vector of ids of the same length as \code{x}.}
}
\value{
A \code{configuration_set}.
}
\description{
\code{configuration_set} creates a \code{configuration_set} object, which is
a list of configurations.
}
\details{
A configuration set is fitted to a one or more group networks. See
\code{\link{fit_group_network}} for details.

\code{set_configuration_ids} creates a \code{configuration_set} and updates the included
configurations' ids.
}
\seealso{
\code{\link{fit_group_network}}, \code{\link{configuration}}
}
